requirements.txt
# ═══════════════════════════════════════════
# Cortana Trading Bot — Dependencies
# ═══════════════════════════════════════════

# Core
pytz==2024.1
httpx==0.27.0
loguru==0.7.2
pydantic==2.7.1
pydantic-settings==2.2.1

# Trading & Finance
alpaca-py>=0.30.1
pandas==2.2.2
pandas-ta==0.3.14b
yfinance==0.2.38
SQLAlchemy==2.0.30

# AI
groq==0.8.0

# Interface
python-telegram-bot>=20.8
APScheduler==3.10.4

# Utils
matplotlib==3.9.0
numpy==1.26.4
tenacity==8.3.0

# Testing
pytest==8.2.1
pytest-asyncio==0.23.7
